refactor(plugin-assistant,plugin-feed): migrate create-object to lazy capabilities
Extract inline create-objects Plugin.addModule blocks from plugin-assistant and
plugin-feed into separate lazy capability files, using AppPlugin.addCreateObjectModule.

- plugin-assistant: create-objects.ts (5 entries incl. Agent) and create-objects-node.ts
  (4 entries without Agent, for the node entry point)
- plugin-feed: create-objects.ts (Feed and Magazine entries)
